% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\docType{data}
\name{adaptive_scales}
\alias{adaptive_scales}
\title{Adaptive Behavior Scales}
\format{
An object of class \code{character} of length 22.
}
\usage{
adaptive_scales
}
\description{
A character vector of scale names related to adaptive behavior skills, daily living abilities, communication skills, socialization, and practical life skills from various adaptive behavior assessment instruments.
}
\keyword{data}
