% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_gt2.R
\name{add_footnotes}
\alias{add_footnotes}
\title{Helper function to add footnotes efficiently}
\usage{
add_footnotes(table, footnote_list)
}
\arguments{
\item{table}{A gt table object}

\item{footnote_list}{List of footnote configurations}
}
\value{
Modified gt table object
}
\description{
Helper function to add footnotes efficiently
}
\keyword{internal}
