% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\docType{data}
\name{adhd_child_scales}
\alias{adhd_child_scales}
\title{ADHD Child Scales}
\format{
An object of class \code{character} of length 44.
}
\usage{
adhd_child_scales
}
\description{
A character vector of scale names related to child and adolescent ADHD symptoms, attention problems, hyperactivity, impulsivity, executive functioning, and emotional regulation from various rating scales and assessment instruments.
}
\keyword{data}
