% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_dependencies}
\alias{check_dependencies}
\title{Utility Functions for Neuropsychological Reports}
\usage{
check_dependencies(packages = NULL)
}
\arguments{
\item{packages}{Character vector of package names}
}
\value{
Logical indicating success
}
\description{
Collection of utility functions used throughout the package.
Check Package Dependencies
}
\details{
Checks if required packages are installed and loads them
}
