% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{create_report_config}
\alias{create_report_config}
\title{Create Report Configuration}
\usage{
create_report_config(
  patient_file,
  template = "adult",
  custom_settings = list()
)
}
\arguments{
\item{patient_file}{Path to patient data}

\item{template}{Report template type}

\item{custom_settings}{List of custom settings}
}
\description{
Create Report Configuration
}
