% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\docType{data}
\name{daily_living_scales}
\alias{daily_living_scales}
\title{Daily Living and Functional Assessment Scales}
\format{
An object of class \code{character} of length 15.
}
\usage{
daily_living_scales
}
\description{
A character vector of scale names related to daily living skills, functional abilities, independent living tasks, and practical life skills from various neuropsychological and functional assessment instruments.
}
\keyword{data}
