% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drilldown2.R
\name{drilldown3}
\alias{drilldown3}
\title{Drilldown on Neuropsychological Domains
This function creates a detailed hierarchical drilldown visualization for cognitive domains.}
\usage{
drilldown3(
  data,
  patient,
  neuro_domain = c("Neuropsychological Test Scores", "Behavioral Rating Scales",
    "Effort/Validity Test Scores"),
  theme
)
}
\arguments{
\item{data}{Dataset to use.}

\item{patient}{Name of patient.}

\item{neuro_domain}{Name of neuropsych domain to add to HC series.}

\item{theme}{The highcharter theme to use.}
}
\value{
A drilldown plot with domain, subdomain, narrow, and scale levels
}
\description{
Drilldown on Neuropsychological Domains
This function creates a detailed hierarchical drilldown visualization for cognitive domains.
}
