% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\docType{data}
\name{emotion_adult_scales}
\alias{emotion_adult_scales}
\title{Adult Emotional and Behavioral Scales}
\format{
An object of class \code{character} of length 85.
}
\usage{
emotion_adult_scales
}
\description{
A character vector of scale names related to adult emotional functioning, mood disorders, personality assessment, behavioral symptoms, and psychological distress from various clinical assessment instruments.
}
\keyword{data}
