% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\docType{data}
\name{emotion_child_scales}
\alias{emotion_child_scales}
\title{Child and Adolescent Emotional and Behavioral Scales}
\format{
An object of class \code{character} of length 85.
}
\usage{
emotion_child_scales
}
\description{
A character vector of scale names related to child and adolescent emotional functioning, mood disorders, behavioral problems, social-emotional development, and psychological symptoms from various clinical assessment instruments.
}
\keyword{data}
