% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_template 2.R, R/domain_template.R,
%   R/domain_template_fixed_filtering.R
\name{generate_qmd_content}
\alias{generate_qmd_content}
\title{Generate the QMD content for a neuropsychological domain}
\usage{
generate_qmd_content(
  domain,
  pheno,
  scales,
  table_filters = NULL,
  exclude_from_plots = NULL,
  subdomain_plot_title,
  narrow_plot_title
)

generate_qmd_content(
  domain,
  pheno,
  scales,
  table_filters = NULL,
  exclude_from_plots = NULL,
  subdomain_plot_title,
  narrow_plot_title
)

generate_qmd_content(
  domain,
  pheno,
  scales,
  table_filters = NULL,
  exclude_from_plots = NULL,
  subdomain_plot_title,
  narrow_plot_title
)
}
\arguments{
\item{domain}{Domain name}

\item{pheno}{Short name for domain}

\item{scales}{Vector of test scales}

\item{table_filters}{Additional filtering criteria}

\item{exclude_from_plots}{Scales to exclude from plots}

\item{subdomain_plot_title}{Caption for subdomain plot}

\item{narrow_plot_title}{Caption for narrow plot}
}
\value{
A character string containing the QMD content

A character string containing the QMD content

A character string containing the QMD content
}
\description{
Generate the QMD content for a neuropsychological domain

Generate the QMD content for a neuropsychological domain

Generate the QMD content for a neuropsychological domain
}
\keyword{internal}
