% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_template 2.R, R/domain_template.R,
%   R/domain_template_fixed_filtering.R
\name{get_default_scales}
\alias{get_default_scales}
\title{Get default scales for a given domain using imported scales from scales.R}
\usage{
get_default_scales(pheno)

get_default_scales(pheno)

get_default_scales(pheno)
}
\arguments{
\item{pheno}{Domain phenotype}
}
\value{
Vector of scale names

Vector of scale names

Vector of scale names
}
\description{
Get default scales for a given domain using imported scales from scales.R

Get default scales for a given domain using imported scales from scales.R

Get default scales for a given domain using imported scales from scales.R
}
\keyword{internal}
