% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_template_fixed.R
\name{get_domain_groupings}
\alias{get_domain_groupings}
\title{Get domain groupings for a specific domain}
\usage{
get_domain_groupings(pheno)
}
\arguments{
\item{pheno}{Domain phenotype}
}
\value{
List of groupings for the domain
}
\description{
Get domain groupings for a specific domain
}
\keyword{internal}
