% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_template.R
\name{get_domain_number}
\alias{get_domain_number}
\title{Get the domain number for a given domain}
\usage{
get_domain_number(pheno)
}
\arguments{
\item{pheno}{Domain phenotype}
}
\value{
Character string with domain number
}
\description{
Get the domain number for a given domain
}
\keyword{internal}
