% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_template_path}
\alias{get_template_path}
\title{Get Template Path}
\usage{
get_template_path(template_name, template_type = "templates")
}
\arguments{
\item{template_name}{Name of the template file}

\item{template_type}{Type of template ('templates', 'assets', etc.)}
}
\value{
Full path to template file
}
\description{
Helper function to get the full path to a template file
}
