% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurotypr.R
\name{merge_lookup_table}
\alias{merge_lookup_table}
\title{Merge Dataset with a Lookup Table}
\usage{
merge_lookup_table(df, lookup_file_path)
}
\arguments{
\item{df}{A data frame to be merged with the lookup table.}

\item{lookup_file_path}{A string specifying the path to the lookup table CSV file.}
}
\value{
A data frame that is the result of merging the input dataset and the lookup table.
}
\description{
This function merges a given dataset with a specified lookup table based on
matching columns. It also ensures the resulting data is sorted and includes
specified missing columns with default values.
}
\examples{
df <- data.frame(
  test = c("Test1", "Test2"),
  test_name = c("Name1", "Name2"),
  scale = c("Scale1", "Scale2")
)
\dontrun{
merge_lookup_table(df, "~/reports/neuropsych_lookup_table_combined.csv")
}
}
