% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\docType{data}
\name{motor_scales}
\alias{motor_scales}
\title{Motor and Psychomotor Scales}
\format{
An object of class \code{character} of length 27.
}
\usage{
motor_scales
}
\description{
A character vector of scale names related to motor functioning, fine motor skills, psychomotor speed, and manual dexterity from various neuropsychological tests.
}
\keyword{data}
