% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurotypr.R
\name{neurotypr_report}
\alias{neurotypr_report}
\title{Create a new Neurotypical Research Report}
\usage{
neurotypr_report()
}
\description{
This function creates a new R Markdown document using the neurotypical research
report template.
}
\examples{
\dontrun{
neurotypr_report()
}
}
