% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurotyp_functions.R
\name{prepare_and_export_spatial_data}
\alias{prepare_and_export_spatial_data}
\title{Prepare and Export Spatial Data}
\usage{
prepare_and_export_spatial_data(
  domains,
  pheno,
  scales,
  input_file = system.file("extdata", "neurocog.csv", package = "NeurotypR"),
  output_file = NULL
)
}
\arguments{
\item{domains}{A character vector of domains to filter the data by.}

\item{pheno}{A string representing the phenotype to use as the base name for the output CSV file.}

\item{scales}{A character vector of scales to filter the data by.}

\item{input_file}{A string representing the name of the input CSV file. Default is "neurocog.csv".}

\item{output_file}{Optional string to specify a custom output file name. If NULL, the file name is derived from \code{pheno}.}
}
\value{
A data frame with the filtered data.
}
\description{
This function reads a CSV file, filters the data based on the specified domains and scales,
selects relevant columns, and writes the filtered data to a new CSV file.
}
\examples{
\dontrun{
domains <- c("Visual Perception/Construction")
pheno <- "spatial"
scales <- c("Arrows", "Bicycle Drawing")
data_spatial <- prepare_and_export_spatial_data(domains, pheno, scales)
}
}
