% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitive_domains.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare Data for Analysis}
\usage{
prepare_data(domains, pheno, scales, type = NULL)
}
\arguments{
\item{domains}{A character vector of domains to filter the data by.}

\item{pheno}{A string representing the phenotype to use as the base name for the output CSV file.}

\item{scales}{A character vector of scales to filter the data by.}

\item{type}{A string representing the name of the input CSV file. Default is "neurocog.csv".}
}
\value{
NULL (invisible) after writing the filtered data to a CSV file.
}
\description{
This function prepares data for analysis by reading a CSV file, filtering the data based on the specified domains and scales,
selecting relevant columns, and writing the filtered data to a new CSV file.
}
\examples{
# Example usage:
domains <- c("Memory")
pheno <- "memory"
scales <- c("Scale1", "Scale2")
prepare_data(domains, pheno, scales)
}
