% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate_row_results.R
\name{process_and_save_results}
\alias{process_and_save_results}
\title{Process and Save Test Results}
\usage{
process_and_save_results(file_path, output_path)
}
\arguments{
\item{file_path}{The path to the CSV file containing test results.}

\item{output_path}{The path where the summarized text file will be saved.}
}
\value{
NULL. Writes the processed summary to the file.
}
\description{
Reads, processes, and generates a concatenated summary for test results,
and writes the summary to a specified file.
}
