\name{run_data_entry_app}
\alias{run_data_entry_app}
\title{Launch Neurotypical Data Entry Shiny App}
\usage{
run_data_entry_app()
}
\value{
A Shiny app object
}
\description{
This function launches a Shiny app for entering and managing neuropsychological data.
The app allows users to input data, save it to a CSV file, and generate reports.
}
\examples{
\dontrun{
run_data_entry_app()
}
}
