% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_template.R
\name{run_neuropsych_domain}
\alias{run_neuropsych_domain}
\title{Run a neuropsych domain through the entire workflow}
\usage{
run_neuropsych_domain(domain_info)
}
\arguments{
\item{domain_info}{List with domain parameters}
}
\value{
Invisibly returns TRUE
}
\description{
Run a neuropsych domain through the entire workflow
}
