% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\docType{data}
\name{scales}
\alias{scales}
\title{All Neuropsychological Test Scales Combined}
\format{
An object of class \code{character} of length 853.
}
\usage{
scales
}
\description{
A comprehensive character vector containing all neuropsychological test scale names from all cognitive and behavioral domains. This is the main scales vector used by the \code{get_neuropsych_scales} function and includes scales from IQ, memory, academics, verbal, spatial, executive, motor, social, ADHD, emotional, adaptive, and daily living domains.
}
\keyword{data}
