% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\docType{data}
\name{scales_academics}
\alias{scales_academics}
\title{Academic Achievement Scales}
\format{
An object of class \code{character} of length 66.
}
\usage{
scales_academics
}
\description{
A character vector of scale names related to academic skills including reading, writing, mathematics, and related cognitive abilities from various academic achievement tests.
}
\keyword{data}
