% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\docType{data}
\name{scales_executive}
\alias{scales_executive}
\title{Executive Function and Attention Scales}
\format{
An object of class \code{character} of length 108.
}
\usage{
scales_executive
}
\description{
A character vector of scale names related to executive functioning, attention, working memory, inhibition, cognitive flexibility, and processing speed from various neuropsychological tests.
}
\keyword{data}
