% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\docType{data}
\name{scales_iq}
\alias{scales_iq}
\title{IQ and General Cognitive Ability Scales}
\format{
An object of class \code{character} of length 39.
}
\usage{
scales_iq
}
\description{
A character vector of scale names related to intelligence quotient (IQ) and general cognitive abilities.
}
\details{
Neuropsychological Test Scales

This file contains character vectors of neuropsychological test scale names,
categorized by cognitive domain. These are used by the
\code{get_neuropsych_scales} function and for defining default scales for
various domains in the \code{create_neuropsych_domain} function.
}
\keyword{data}
