% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\docType{data}
\name{scales_memory}
\alias{scales_memory}
\title{Memory Scales}
\format{
An object of class \code{character} of length 135.
}
\usage{
scales_memory
}
\description{
A character vector of scale names related to memory functions, including recall, recognition, and learning, from various tests like CVLT, NAB, etc.
}
\keyword{data}
