% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\docType{data}
\name{scales_spatial}
\alias{scales_spatial}
\title{Spatial and Visual-Constructional Scales}
\format{
An object of class \code{character} of length 42.
}
\usage{
scales_spatial
}
\description{
A character vector of scale names related to spatial abilities, visual-perceptual skills, and constructional tasks from various neuropsychological tests.
}
\keyword{data}
