% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\docType{data}
\name{scales_verbal}
\alias{scales_verbal}
\title{Verbal and Language Scales}
\format{
An object of class \code{character} of length 80.
}
\usage{
scales_verbal
}
\description{
A character vector of scale names related to verbal abilities, language comprehension, expression, and communication skills from various neuropsychological and language tests.
}
\keyword{data}
