% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\docType{data}
\name{social_scales}
\alias{social_scales}
\title{Social Cognition and Autism Spectrum Scales}
\format{
An object of class \code{character} of length 57.
}
\usage{
social_scales
}
\description{
A character vector of scale names related to social cognition, theory of mind, affect recognition, autism spectrum characteristics, and social interaction skills from various neuropsychological and developmental tests.
}
\keyword{data}
