% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_render_typ.R
\name{summary_function}
\alias{summary_function}
\title{Summary Function}
\usage{
summary_function(files)
}
\arguments{
\item{files}{A vector of file names}
}
\value{
A list of summaries, one for each file
}
\description{
Reads in a vector of file names and returns a summary of each file
}
