% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_data_structure}
\alias{validate_data_structure}
\title{Validate Data Structure}
\usage{
validate_data_structure(data)
}
\arguments{
\item{data}{Data frame or list to validate}
}
\value{
Logical indicating if data is valid
}
\description{
Validates that the input data has the expected structure for processing
}
