% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitive_domains.R
\name{visualize_data}
\alias{visualize_data}
\title{Visualize Data with Table and Dot Plot}
\usage{
visualize_data(
  data,
  scales_to_keep,
  pheno,
  domain,
  grp_pheno = NULL,
  x = "x",
  y = "y"
)
}
\arguments{
\item{data}{A data frame containing the data to be visualized. Must contain columns \code{scale}, \code{x}, and \code{y}.}

\item{scales_to_keep}{A vector of scales to filter the data by.}

\item{pheno}{A string representing the phenotype of interest.}

\item{domain}{A string representing the domain of interest.}

\item{grp_pheno}{A list containing grouped phenotypic data. Default is NULL.}

\item{x}{A string representing the name of the x variable. This should match a column name in the data.}

\item{y}{A string representing the name of the y variable. This should match a column name in the data.}
}
\value{
A list containing two elements: the table visualization (from \code{tbl_gt}) and the dot plot visualization (from \code{dotplot}).
}
\description{
This function generates a table and a dot plot for a given dataset based on specified phenotypes and domains.
The table is created using \code{tbl_gt}, and the dot plot is created using \code{dotplot}.
}
\examples{
# Example usage:
data <- data.frame(scale = c("A", "B", "C"), x = rnorm(3), y = rnorm(3))
scales_to_keep <- c("A", "B")
pheno <- "example_pheno"
domain <- "example_domain"
visualize_data(data, scales_to_keep, pheno, domain)
}
