% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_template.R
\name{get_default_scales}
\alias{get_default_scales}
\title{Get default scales for a given domain}
\usage{
get_default_scales(pheno)
}
\arguments{
\item{pheno}{Domain phenotype}
}
\value{
Vector of scale names
}
\description{
Get default scales for a given domain
}
\keyword{internal}
