% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_template 2.R, R/domain_template.R,
%   R/domain_template_fixed_filtering.R
\name{get_domain_number}
\alias{get_domain_number}
\title{Get the domain number for a given domain}
\usage{
get_domain_number(pheno)

get_domain_number(pheno)

get_domain_number(pheno)
}
\arguments{
\item{pheno}{Domain phenotype}
}
\value{
Character string with domain number

Character string with domain number

Character string with domain number
}
\description{
Get the domain number for a given domain

Get the domain number for a given domain

Get the domain number for a given domain
}
\keyword{internal}
